Go-service deployment
Deploys three docker containers, one with mongodb, one with a rest-api
built using Python EVE and one which runs the redirection service.

A docker network is setup between the two containers which enables them
to communicate.

Use http://<Service-IP>:12001/login to get a token.

Using http://<Service-IP>:12001/api/all one is able to list all current
aliases.

Using http://<Service-IP>:12001/api/all/<ldapuser> one is able to list all
current aliases for a specific user.

Using http://<Service-IP>:12001/api one is able to post a new
pattern and alias pair. The pattern can be a regex.

Using http://<Service-IP>:12001/api one is able to update a alias.

Using http:<Service-IP>:12000/<alias> it either redirects or casts 404
depending if the alias is present or not.

Using http:<Service-IP>:12001/api/<id> will
remove the corresponding alias. Only removes if both pattern and owner
are correct.

If a user is part of one of the admin groups, then that user is allowed
to remove and alter any entry.

Change-Id: If4ec3bb8ae6d5e406a6cfe76878dacc899d98f33