Fix pymongo warning while using multiple threads
As UWSGI uses multiple threads while serving our flask
applications it is recommended to handle the database
connection in a specific way to reduce the load during
multiple quick restarts. Setting `lazy-apps = true`
makes sure that UWSGI forks the application before the
database connection is established.

The api application cannot use this approch due to
the the authentication needing one single truth and
cannot be split.

Change-Id: I1afeab08c318767880d9fd904bf5f0c6f81f3135