Perform major upgrade before deployment
Change the ansible deployment scripts to a simplfied docker-compose
structure. Currently the structure is composed of four containers.

The database container is upgraded to mongoDB 3.6 and given persistent
volume and initial authentication if needed. The port is changed from
default to 27028 for increased security.

The api container is now build from official python containers instead
of unofficial premade uwsgi-nginx. Nginx is moved to a seperate
container and uwsgi is no longer run as root. A requirement.txt file
is added for clearity.

The service container is also build from offical python containers
without the nginx dependency. The settings.py file is removed for
simpler config directly inside service.py. Uwsgi is no longer run
as root and lazy-loaded for compliance with pymongo forking.

The nginx container is build from offical nginx image with certificates
mounted in read only mode. This container is the only one exposed to
external networks and only has access to the api and service containers
not the database container. This container serves the static site on
root path ("go/") and redirects api requests incoming at /api to the
api container via uwsgi sockets unknown paths are directed to the
service container via the same uwsgi socket configuration. The nginx
config tries to follow common best-practices to improve performance.

A bunch of other minor changes were needed to fix bugs improve the ssl
configuration or clean up the code base.
-----------------------------------------------------------+-----------+---

Change-Id: Icf963e0d1fdfa7ed79550e65d7854a4517e49c7a