Change to reverse regex match database lookup
In order to accommodate previous functionality we needed to change the
query method. This means that for every request the database needs to
go through the entire list of entries and check for regex match.
Potentially greatly reducing performance. However, since the
functionality is needed and the number of entries is considered small
this will hopefully never be a problem.

Change-Id: I5b294ba32553340ba6be2918e6b49768090f8d0b