Fix less naive ranking of pattern that match alias
Previously we simply returned the target of the longest pattern that
matched the incoming alias. This rewrite takes a less naive approach and
utilizes built in functionality in pythons regex engine.

In the sre_parse module which is used by the re module there is a
parse method that converts the regex string into a sre_parse.SubPattern
object. This object is basically a tree like structure of how the regex
is interpreted. This object has a method called getwidth() which returns
a tuple of the minimum and maximum width of the parsed regex. This is a
much better measure to sort the regex by than by length.

With this implementation we return the pattern with lowest width first
and if multiple exist we return to shortest pattern (unless infinite
wildcards exist then longest pattern) and finally if they are of the same
length we revert to alphabetical order.

This approach does not break any of the existing shortcuts and provides a
more intuitive approach to the ranking. For example "(.*)" now has a
much lower ranking than "atf", "a", "(atf|atf-tool)" as is expected. It
also has the benefit of ranking "atf" higher than "at\w" and "atf"
higher than "(atf)" and "(a|atf)".

Some minor issues may remain which may require more unit tests.

Change-Id: I6bc97ed34999306530e256e2a2cc304fbbaacc89