Add automatic development environment setup
Utilizing an additional docker-compose file and some minor changes to
the configuration for a automatic setup of a development environment.

This environment is invoced by running
`docker-compose -f docker-compose.dev.yml up --build`

This starts the backend containers with flask development server instead
of the production uwsgi server and with debugging option turned on. The
source files are volume mounted which enables automatic restart of the
applications on every file change without the need to rebuild the images.

The frontend is not brought up with docker in the debugging environment
instead this is done by the usual `yarn start` while standing in the
frontend folder. However, it detects a development environment
automatically and sets the api url correctly.

Change-Id: I6d43ea989aad7dd70215428535f630872f338bfb