Refactor Dockerfiles to reduce size and build steps
The nginx container now utilizes the `node:9-alpine` image which is half
the size of the `node:9-slim image`. We also clear the yarn cache during
container build to further reduce the size.

For the other dockerfiles we reduce the build steps by merging multiple
COPY commands into a single command.

Change-Id: I36c71c25bd0069aa35cb01d009bf74f512146b31