Fix broken styling on sort arrows
Due to broken behavior in the latest version of react-icons-kit (1.1.6)
the arrows used to indicate sorting took up more space on the page. The
version published to NPM is not available to github yet but I wrote an
issue to the author of the package.

This fix pins the version to the latest working release (1.1.5).

Change-Id: Ief37a90aa2004fffe2a9d7f5e7ecc4e9d2e37464