Change to improved font and cache policy
Sets the expires header to one year on static files. New files are still
downloaded on changes due to the etag header.

Fix the mime type of woff/woff2 font files and disables gzip on these
files since the format is already compressed.

Sets the font-display property to block instead of auto to fix
renderinging of invisible text when font response is slow.

Change-Id: I10fcec732e3186af4c082c6d1720535b7a11bfe0