Fix: only append email domain to user if configured
Some LDAP setups might not have logins where the user is configured as
the email address but instead only the username. Such setups will
specify `LDAP_EMAIL_DOMAIN` as blank. That would not work before since
we would still append the `@` symbol on the end.

This change makes sure to only append the `@` symbol of the email domain
is not blank in the configuration.