Add additional pattern validation
Add validation of pattern to make sure that 'api' and patterns starting
with 'api/' is not valid. This is to make sure that users do not insert
patterns which would not work due to being in use by the backend api and
not going towards the redirect service at all.