Fix issues with special characters in search bar
If the filter string from the search bar included special characters
but still valid regex strings it was not escaped correctly which lead to
errors down the line in the front-end code when a json request string
was being constructed.

To fix this I simply escape such problematic characters before sending
the string down-stream to execute the filtering. This has the benefit of
behaving close to how users would expect. A search for \1 results in a
request towards mongodb with \\1 which returns strings with \1 in them.